// src/Storefront.WebApp/app/src/workplace/enums/agreement/agreementClassifierTypeCode.enum.ts
export enum AgreementClassifierTypeCodeEnum {
  InactiveAgreement = 'INACTIVE_AGREEMENT',
  BlockLimitArrest = 'BLOCK_LIMIT_ARREST',
  BlockFullArrest = 'BLOCK_FULL_ARREST',
  BlockFmDebit = 'BLOCK_FM_DEBIT',
  BlockFmCredit = 'BLOCK_FM_CREDIT',
  BlockFmFull = 'BLOCK_FM_FULL',
  BlockFmIdentification = 'BLOCK_FM_IDENTIFICATION',
  BlockDecisionBank = 'BLOCK_DECISION_BANK',
  BlockGrodi = 'BLOCK_GRODI',
  BlockGna = 'BLOCK_GNA',
  BlockTerrorist = 'BLOCK_TERRORIST',
  BlockPep = 'BLOCK_PEP',
  BlockDeathClient = 'BLOCK_DEATH_CLIENT',
  BlockOverdueLoan = 'BLOCK_OVERDUE_LOAN',
  BlockCustom = 'BLOCK_CUSTOM',
  BlockResidentStatusChange = 'BLOCK_RESIDENT_STATUS_CHANGE',
  BlockClientMissinLists = 'BLOCK_CLIENT_MISSIN_LISTS',
  BlockInvalidPassport = 'BLOCK_INVALID_PASSPORT',
  Undefined = 'Undefined'
}